+++ CHANGELOG.rst
@@ -6,6 +6,7 @@ Version 0.1.12 - 2021-MM-DD
 ===========================
 
 - New source: suntime
+- New sink: Tuya/Smart Life
 
 Version 0.1.11 - 2021-05-05
 ===========================